multiplication result converted to larger type

+++ MagickCore/statistic.c
@@ -2255,15 +2255,16 @@ MagickExport ChannelStatistics *GetImageStatistics(const Image *image,
     area=PerceptibleReciprocalLD(channel_statistics[channel].area);
     for (j=0; j <= (ssize_t) MaxMap; j++)
     {
-      long double
+      double
         count;
 
       count=area*histogram[(ssize_t) GetPixelChannels(image)*j+i];
-      channel_statistics[channel].entropy+=(double) (-count*MagickLog10(count)*
-        PerceptibleReciprocalLD(MagickLog10(number_bins)));
-      channel_statistics[CompositePixelChannel].entropy+=(double) (-count*
-        MagickLog10(count)*PerceptibleReciprocalLD(MagickLog10(number_bins))/
-        GetPixelChannels(image));
+      channel_statistics[channel].entropy+=((long double) -count*
+        MagickLog10(count)*PerceptibleReciprocalLD((long double)
+        MagickLog10(number_bins)));
+      channel_statistics[CompositePixelChannel].entropy+=((long double) -count*
+        MagickLog10(count)*PerceptibleReciprocalLD((long double)
+        MagickLog10(number_bins))/GetPixelChannels(image));
     }
   }
   histogram=(double *) RelinquishMagickMemory(histogram);
